+++ swan/dataset/__init__.py
@@ -1,5 +1,6 @@
-from .fingerprints_data import FingerprintsData
-from .torch_geometric_graph_data import TorchGeometricGraphData
 from .dgl_graph_data import DGLGraphData
+from .fingerprints_data import FingerprintsData
+from .splitter import split_dataset, load_split_dataset
+from .torch_geometric_graph_data import TorchGeometricGraphData
 
-__all__ = ["DGLGraphData", "FingerprintsData", "TorchGeometricGraphData"]
+__all__ = ["DGLGraphData", "FingerprintsData", "TorchGeometricGraphData", "load_split_dataset", "split_dataset"]