+++ tests/test_dataset.py
@@ -8,25 +8,25 @@ PATH_CSV = PATH_TEST / "thousand.csv"
 
 def test_fingerprint_dataset():
     """Check that the fingerprint dataset is loaded correctly."""
-    data = FingerprintsData(PATH_CSV, properties=["gammas"])
+    data = FingerprintsData(PATH_CSV, properties=["Hardness (eta)"])
     data.create_data_loader()
 
 
 def test_torch_geometric_dataset():
     """Check that the torch_geometric dataset is loaded correctly."""
-    data = TorchGeometricGraphData(PATH_CSV, properties=["gammas"])
+    data = TorchGeometricGraphData(PATH_CSV, properties=["Hardness (eta)"])
     data.create_data_loader()
 
 
 def test_torch_geometric_dataset_with_optimization():
     """Check that the torch_geometric dataset generates a guess for the molecular coordinate."""
-    data = TorchGeometricGraphData(PATH_CSV, properties=["gammas"], optimize_molecule=True)
+    data = TorchGeometricGraphData(PATH_CSV, properties=["Hardness (eta)"], optimize_molecule=True)
     data.create_data_loader()
 
 
 def test_dgl_dataset():
     """Check that the DGL dataset is loaded correctly."""
-    data = DGLGraphData(PATH_CSV, properties=["gammas"])
+    data = DGLGraphData(PATH_CSV, properties=["Hardness (eta)"])
     data.create_data_loader()