+++ tests/test_fingerprint_modeller.py
@@ -1,7 +1,7 @@
 import unittest
 import numpy as np
 import torch
-from swan.modeller import Modeller
+from swan.modeller import TorchModeller
 from swan.modeller.models.fingerprint_models import FingerprintFullyConnected
 from swan.dataset import FingerprintsData
 
@@ -11,16 +11,15 @@ from .utils_test import PATH_TEST, remove_files
 class TestFingerprintModeller(unittest.TestCase):
     """Test the finger print models"""
     def setUp(self):
-        data = FingerprintsData(PATH_TEST / "thousand.csv", properties=["gammas"])
+        data = FingerprintsData(PATH_TEST / "thousand.csv", properties=["Hardness (eta)"])
         self.net = FingerprintFullyConnected()
-        self.modeller = Modeller(self.net, data)
+        self.modeller = TorchModeller(self.net, data)
 
     def test_train(self):
         self.modeller.data.scale_labels()
         self.modeller.train_model(nepoch=5, batch_size=64)
         expected, predicted = self.modeller.validate_model()
-        err = torch.functional.F.mse_loss(expected, predicted)
-        assert not np.isnan(err.item())
+        assert not all(np.isnan(x).all() for x in (expected, predicted))
         remove_files()
 
     def test_predict(self):