+++ tests/test_mpnn.py
@@ -5,7 +5,7 @@ import torch
 import torch_geometric
 
 from swan.dataset import TorchGeometricGraphData
-from swan.modeller import Modeller
+from swan.modeller import TorchModeller
 from swan.modeller.models.graph_models import MPNN
 
 from .utils_test import PATH_TEST, remove_files
@@ -16,17 +16,16 @@ class TestMPNN(unittest.TestCase):
     """
     def setUp(self):
         self.data = PATH_TEST / "thousand.csv"
-        self.data = TorchGeometricGraphData(self.data, properties=["gammas"])
+        self.data = TorchGeometricGraphData(self.data, properties=["Hardness (eta)"])
         self.net = MPNN()
-        self.modeller = Modeller(self.net, self.data)
+        self.modeller = TorchModeller(self.net, self.data)
 
     def test_train_data_mpnn(self):
 
         self.modeller.data.scale_labels()
         self.modeller.train_model(nepoch=5, batch_size=64)
         expected, predicted = self.modeller.validate_model()
-        err = torch.functional.F.mse_loss(expected, predicted)
-        assert not np.isnan(err.item())
+        assert not all(np.isnan(x).all() for x in (expected, predicted))
         remove_files()
 
     def test_predict(self):