+++ tests/test_scikit_models.py
@@ -7,13 +7,13 @@ from swan.modeller import SKModeller
 
 from .utils_test import PATH_TEST
 
-DATA = FingerprintsData(PATH_TEST / "thousand.csv", properties=["gammas"], sanitize=False)
+DATA = FingerprintsData(PATH_TEST / "thousand.csv", properties=["Hardness (eta)"], sanitize=False)
 DATA.scale_labels()
 
 
 def run_test(model: str, **kwargs):
     """Run the training and validation step for the given model."""
-    modeller = SKModeller(DATA, model)
+    modeller = SKModeller(model, DATA)
     modeller.train_model()
     predicted, expected = modeller.validate_model()
     reg = stats.linregress(predicted.flatten(), expected.flatten())