+++ tests/test_se3_transformer.py
@@ -3,7 +3,7 @@ import torch
 
 from swan.dataset import DGLGraphData
 from swan.dataset.dgl_graph_data import dgl_data_loader
-from swan.modeller import Modeller
+from swan.modeller import TorchModeller
 from swan.modeller.models import TFN, SE3Transformer
 
 from .utils_test import PATH_TEST, remove_files
@@ -15,18 +15,17 @@ NUM_CHANNELS = 4  # Number of channels in middle layers
 torch.set_default_dtype(torch.float32)
 
 CSV_FILE = PATH_TEST / "thousand.csv"
-DATA = DGLGraphData(CSV_FILE, properties=["gammas"])
+DATA = DGLGraphData(CSV_FILE, properties=["Hardness (eta)"])
 
 
 def run_modeller(net: torch.nn.Module):
     """Run a given model."""
-    modeller = Modeller(net, DATA, use_cuda=False, replace_state=False)
+    modeller = TorchModeller(net, DATA, use_cuda=False, replace_state=False)
 
     modeller.data.scale_labels()
     modeller.train_model(nepoch=1, batch_size=64)
     expected, predicted = modeller.validate_model()
-    err = torch.functional.F.mse_loss(expected, predicted)
-    assert not np.isnan(err.item())
+    assert not all(np.isnan(x).all() for x in (expected, predicted))
     remove_files()
 
 
@@ -45,7 +44,7 @@ def test_s3eTransformer_train():
 def test_se3Transformer_predict():
     """Check the prediction functionality of the SE3Transformer."""
     net = SE3Transformer(NUM_LAYERS, NUM_CHANNELS)
-    researcher = Modeller(net, DATA, use_cuda=False)
+    researcher = TorchModeller(net, DATA, use_cuda=False)
     researcher.load_model("swan_chk.pt")
 
     # Predict the properties